+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [1.0.0] - TBD
+
+## [1.0.0+pre1] - 2025-06-08
+
+This is our first prerelease of version [1.0.0](#100---tbd). Also because of the way we committed code last time [0.1.3](#013snapshot25w24a---2025-06-08)'s code was a little corrupt.
+
+### Changed
+
+- Changed the name of the project to muddybot
+
+### Added
+
+- Added some error handling for when users don't send arguments with commands that require arguments
+
+### Removed
+
+- Removed !wordleonwheels command
+- Removed unused except `KeyboardInterrupt:` code block
+- Some unused classifiers in pyproject.toml
+
 ## [0.1.3+snapshot25w24a] - 2025-06-08
 
 ### Changed
@@ -67,3 +87,4 @@ _Initial release_
 [0.1.1+snapshot25w23a]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.1+snapshot25w23a
 [0.1.2+snapshot25w23b]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.2+snapshot25w23b
 [0.1.3+snapshot25w24a]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.3+snapshot25w24a
+[1.0.0+pre1]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.0+pre1